+++ tests/testthat/test-repositories.R
@@ -5,7 +5,20 @@ context("repositories")
 
 suffix <- sample(letters, 10, replace = TRUE) %>% str_c(collapse = "")
 
+setup(suppressMessages({
+
+  create_team(
+    name        = str_c("test-repositories-", suffix),
+    org         = "HairyCoos",
+    description = "This is a team to test repositories")
+
+}))
+
 teardown(suppressMessages({
+
+  try(silent = TRUE, {
+    delete_team(str_c("test-repositories-", suffix), org = "HairyCoos")
+  })
 
   try(silent = TRUE, {
     delete_repository(str_c("ChadGoymer/user-repository-", suffix))
@@ -222,7 +235,7 @@ test_that("update_repository changes a repository's properties and returns them 
   expect_true(org_repo$allow_rebase_merge)
 
 
-  archived_repo <- update_repository(str_c("HairyCoos/updated-org-repository-", suffix), archived = TRUE)
+  archived_repo <- update_repository(str_c("ChadGoymer/updated-user-repository-", suffix), archived = TRUE)
 
   expect_is(archived_repo, "list")
   expect_identical(attr(archived_repo, "status"), 200L)
@@ -256,6 +269,47 @@ test_that("update_repository changes a repository's properties and returns them 
       updated_at         = "POSIXct"))
 
   expect_true(archived_repo$archived)
+
+})
+
+
+# TEST: update_team_repository ----------------------------------------------------------------
+
+test_that("update_team_repository adds and updates a teams permission to a repository", {
+
+  added_repo <- update_team_repository(
+    repo       = str_c("HairyCoos/updated-org-repository-", suffix),
+    team       = str_c("test-repositories-", suffix),
+    org        = "HairyCoos",
+    permission = "pull")
+
+  expect_is(added_repo, "logical")
+  expect_identical(attr(added_repo, "status"), 204L)
+  expect_identical(as.logical(added_repo), TRUE)
+
+  read_repo <- view_repository(
+    repo = str_c("HairyCoos/updated-org-repository-", suffix),
+    team = str_c("test-repositories-", suffix),
+    org  = "HairyCoos")
+
+  expect_identical(read_repo$permission, "pull")
+
+  updated_repo <- update_team_repository(
+    repo       = str_c("HairyCoos/updated-org-repository-", suffix),
+    team       = str_c("test-repositories-", suffix),
+    org        = "HairyCoos",
+    permission = "maintain")
+
+  expect_is(updated_repo, "logical")
+  expect_identical(attr(updated_repo, "status"), 204L)
+  expect_identical(as.logical(updated_repo), TRUE)
+
+  maintain_repo <- view_repository(
+    repo = str_c("HairyCoos/updated-org-repository-", suffix),
+    team = str_c("test-repositories-", suffix),
+    org  = "HairyCoos")
+
+  expect_identical(maintain_repo$permission, "maintain")
 
 })
 
@@ -375,6 +429,51 @@ test_that("view_repositories returns a tibble summarising the repositories", {
       updated_at         = "POSIXct"))
 
   expect_true(str_c("updated-org-repository-", suffix) %in% org_repos$name)
+  expect_identical(
+    org_repos %>% filter(name == str_c("updated-org-repository-", suffix)) %>% pull("permission"),
+    "admin")
+
+
+  team_repos <- view_repositories(
+    team  = str_c("test-repositories-", suffix),
+    org   = "HairyCoos",
+    n_max = 10)
+
+  expect_is(team_repos, "tbl")
+  expect_identical(attr(team_repos, "status"), 200L)
+  expect_identical(
+    map_chr(team_repos, ~ class(.)[[1]]),
+    c(id                 = "integer",
+      name               = "character",
+      full_name          = "character",
+      description        = "character",
+      owner              = "character",
+      homepage           = "character",
+      language           = "character",
+      size               = "numeric",
+      default_branch     = "character",
+      permission         = "character",
+      private            = "logical",
+      has_issues         = "logical",
+      has_projects       = "logical",
+      has_wiki           = "logical",
+      has_pages          = "logical",
+      has_downloads      = "logical",
+      allow_squash_merge = "logical",
+      allow_merge_commit = "logical",
+      allow_rebase_merge = "logical",
+      fork               = "logical",
+      archived           = "logical",
+      disabled           = "logical",
+      html_url           = "character",
+      pushed_at          = "POSIXct",
+      created_at         = "POSIXct",
+      updated_at         = "POSIXct"))
+
+  expect_true(str_c("updated-org-repository-", suffix) %in% team_repos$name)
+  expect_identical(
+    team_repos %>% filter(name == str_c("updated-org-repository-", suffix)) %>% pull("permission"),
+    "maintain")
 
 
   auth_repos <- view_repositories(n_max = 10)
@@ -468,6 +567,29 @@ test_that("browse_repository opens the repository's page in the browser", {
   expect_is(repo, "character")
   expect_identical(attr(repo, "status"), 200L)
   expect_identical(as.character(repo), str_c("https://github.com/ChadGoymer/updated-user-repository-", suffix))
+
+})
+
+
+# TEST: remove_team_repository ----------------------------------------------------------------
+
+test_that("remove_team_repository removes a teams acces to a repository", {
+
+  removed_repo <- remove_team_repository(
+    repo       = str_c("HairyCoos/updated-org-repository-", suffix),
+    team       = str_c("test-repositories-", suffix),
+    org        = "HairyCoos")
+
+  expect_is(removed_repo, "logical")
+  expect_identical(attr(removed_repo, "status"), 204L)
+  expect_identical(as.logical(removed_repo), TRUE)
+
+  team_repos <- view_repositories(
+    team  = str_c("test-repositories-", suffix),
+    org   = "HairyCoos",
+    n_max = 10)
+
+  expect_identical(nrow(team_repos), 0L)
 
 })